+++ api/spec/controllers/spree/api/checkouts_controller_spec.rb
@@ -38,21 +38,21 @@ module Spree
       it "should transition a recently created order from cart to address" do
         order.state.should eq "cart"
         order.email.should_not be_nil
-        api_put :update, :id => order.to_param, :order_token => order.token
+        api_put :update, :id => order.to_param, :order_token => order.guest_token
         order.reload.state.should eq "address"
       end
 
       it "should transition a recently created order from cart to address with order token in header" do
         order.state.should eq "cart"
         order.email.should_not be_nil
-        request.headers["X-Spree-Order-Token"] = order.token
+        request.headers["X-Spree-Order-Token"] = order.guest_token
         api_put :update, :id => order.to_param
         order.reload.state.should eq "address"
       end
 
       it "can take line_items_attributes as a parameter" do
         line_item = order.line_items.first
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
                          :order => { :line_items_attributes => { 0 => { :id => line_item.id, :quantity => 1 } } }
         response.status.should == 200
         order.reload.state.should eq "address"
@@ -60,7 +60,7 @@ module Spree
 
       it "can take line_items as a parameter" do
         line_item = order.line_items.first
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
                          :order => { :line_items => { 0 => { :id => line_item.id, :quantity => 1 } } }
         response.status.should == 200
         order.reload.state.should eq "address"
@@ -71,7 +71,7 @@ module Spree
         order.bill_address = nil
         order.save
         order.update_column(:state, "address")
-        api_put :update, :id => order.to_param, :order_token => order.token
+        api_put :update, :id => order.to_param, :order_token => order.guest_token
         # Order has not transitioned
         response.status.should == 422
       end
@@ -96,7 +96,7 @@ module Spree
 
         it "can update addresses and transition from address to delivery" do
           api_put :update,
-                  :id => order.to_param, :order_token => order.token,
+                  :id => order.to_param, :order_token => order.guest_token,
                   :order => {
                     :bill_address_attributes => address,
                     :ship_address_attributes => address
@@ -110,7 +110,7 @@ module Spree
         # Regression test for #4498
         it "does not contain duplicate variant data in delivery return" do
           api_put :update,
-                  :id => order.to_param, :order_token => order.token,
+                  :id => order.to_param, :order_token => order.guest_token,
                   :order => {
                     :bill_address_attributes => address,
                     :ship_address_attributes => address
@@ -127,7 +127,7 @@ module Spree
         shipment = create(:shipment, :order => order)
         shipment.refresh_rates
         shipping_rate = shipment.shipping_rates.where(:selected => false).first
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
           :order => { :shipments_attributes => { "0" => { :selected_shipping_rate_id => shipping_rate.id, :id => shipment.id } } }
         response.status.should == 200
         # Find the correct shipment...
@@ -142,7 +142,7 @@ module Spree
 
       it "can update payment method and transition from payment to confirm" do
         order.update_column(:state, "payment")
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
           :order => { :payments_attributes => [{ :payment_method_id => @payment_method.id }] }
         json_response['state'].should == 'confirm'
         json_response['payments'][0]['payment_method']['name'].should == @payment_method.name
@@ -160,7 +160,7 @@ module Spree
           "name" => "Spree Commerce"
         }
 
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
           :order => { :payments_attributes => [{ :payment_method_id => @payment_method.id.to_s }],
                       :payment_source => { @payment_method.id.to_s => source_attributes } }
         json_response['payments'][0]['payment_method']['name'].should == @payment_method.name
@@ -170,7 +170,7 @@ module Spree
 
       it "returns errors when source is missing attributes" do
         order.update_column(:state, "payment")
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
           :order => {
             :payments_attributes => [{ :payment_method_id => @payment_method.id.to_s }]
           },
@@ -189,14 +189,14 @@ module Spree
       it "can transition from confirm to complete" do
         order.update_column(:state, "confirm")
         Spree::Order.any_instance.stub(:payment_required? => false)
-        api_put :update, :id => order.to_param, :order_token => order.token
+        api_put :update, :id => order.to_param, :order_token => order.guest_token
         json_response['state'].should == 'complete'
         response.status.should == 200
       end
 
       it "returns the order if the order is already complete" do
         order.update_column(:state, "complete")
-        api_put :update, :id => order.to_param, :order_token => order.token
+        api_put :update, :id => order.to_param, :order_token => order.guest_token
         json_response['number'].should == order.number
         response.status.should == 200
       end
@@ -204,7 +204,7 @@ module Spree
       # Regression test for #3784
       it "can update the special instructions for an order" do
         instructions = "Don't drop it. (Please)"
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
           :order => { :special_instructions => instructions }
         expect(json_response['special_instructions']).to eql(instructions)
       end
@@ -214,7 +214,7 @@ module Spree
         it "can assign a user to the order" do
           user = create(:user)
           # Need to pass email as well so that validations succeed
-          api_put :update, :id => order.to_param, :order_token => order.token,
+          api_put :update, :id => order.to_param, :order_token => order.guest_token,
             :order => { :user_id => user.id, :email => "guest@spreecommerce.com" }
           response.status.should == 200
           json_response['user_id'].should == user.id
@@ -222,7 +222,7 @@ module Spree
       end
 
       it "can assign an email to the order" do
-        api_put :update, :id => order.to_param, :order_token => order.token,
+        api_put :update, :id => order.to_param, :order_token => order.guest_token,
           :order => { :email => "guest@spreecommerce.com" }
         json_response['email'].should == "guest@spreecommerce.com"
         response.status.should == 200
@@ -234,7 +234,7 @@ module Spree
         order.update_column(:state, "payment")
         PromotionHandler::Coupon.should_receive(:new).with(order).and_call_original
         PromotionHandler::Coupon.any_instance.should_receive(:apply).and_return({:coupon_applied? => true})
-        api_put :update, :id => order.to_param, :order_token => order.token, :order => { :coupon_code => "foobar" }
+        api_put :update, :id => order.to_param, :order_token => order.guest_token, :order => { :coupon_code => "foobar" }
       end
     end
 
@@ -243,7 +243,7 @@ module Spree
       it "cannot transition to address without a line item" do
         order.line_items.delete_all
         order.update_column(:email, "spree@example.com")
-        api_put :next, :id => order.to_param, :order_token => order.token
+        api_put :next, :id => order.to_param, :order_token => order.guest_token
         response.status.should == 422
         json_response["errors"]["base"].should include(Spree.t(:there_are_no_items_for_this_order))
       end
@@ -251,7 +251,7 @@ module Spree
       it "can transition an order to the next state" do
         order.update_column(:email, "spree@example.com")
 
-        api_put :next, :id => order.to_param, :order_token => order.token
+        api_put :next, :id => order.to_param, :order_token => order.guest_token
         response.status.should == 200
         json_response['state'].should == 'address'
       end
@@ -262,14 +262,14 @@ module Spree
           email: nil
         )
 
-        api_put :next, :id => order.to_param, :order_token => order.token
+        api_put :next, :id => order.to_param, :order_token => order.guest_token
         response.status.should == 422
         json_response['error'].should =~ /could not be transitioned/
       end
 
       it "doesnt advance payment state if order has no payment" do
         order.update_column(:state, "payment")
-        api_put :next, :id => order.to_param, :order_token => order.token, :order => {}
+        api_put :next, :id => order.to_param, :order_token => order.guest_token, :order => {}
         json_response["errors"]["base"].should include(Spree.t(:no_payment_found))
       end
     end
@@ -279,11 +279,11 @@ module Spree
 
       it 'continues to advance advances an order while it can move forward' do
         Spree::Order.any_instance.should_receive(:next).exactly(3).times.and_return(true, true, false)
-        api_put :advance, :id => order.to_param, :order_token => order.token
+        api_put :advance, :id => order.to_param, :order_token => order.guest_token
       end
 
       it 'returns the order' do
-        api_put :advance, :id => order.to_param, :order_token => order.token
+        api_put :advance, :id => order.to_param, :order_token => order.guest_token
         json_response['id'].should == order.id
       end
     end