+++ core/app/models/spree/order.rb
@@ -637,7 +637,10 @@ module Spree
       end
 
       def create_token
-        self.token = ::SecureRandom::hex(8)
+        self.guest_token ||= loop do
+          random_token = SecureRandom.urlsafe_base64(nil, false)
+          break random_token unless self.class.exists?(guest_token: random_token)
+        end
       end
   end
 end