+++ core/db/migrate/20140530024945_move_order_token_from_tokenized_permission.rb
@@ -8,16 +8,16 @@ class MoveOrderTokenFromTokenizedPermission < ActiveRecord::Migration
     when 'SQLite'
       Spree::Order.has_one :tokenized_permission, :as => :permissable
       Spree::Order.includes(:tokenized_permission).each do |o|
-        o.update_column :token, o.tokenized_permission.token
+        o.update_column :guest_token, o.tokenized_permission.token
       end
     when 'Mysql2'
       execute "UPDATE spree_orders, spree_tokenized_permissions
-               SET spree_orders.token = spree_tokenized_permissions.token
+               SET spree_orders.guest_token = spree_tokenized_permissions.token
                WHERE spree_tokenized_permissions.permissable_id = spree_orders.id
                   AND spree_tokenized_permissions.permissable_type = 'Spree::Order'"
     else
       execute "UPDATE spree_orders
-               SET token = spree_tokenized_permissions.token
+               SET guest_token = spree_tokenized_permissions.token
                FROM spree_tokenized_permissions
                WHERE spree_tokenized_permissions.permissable_id = spree_orders.id
                   AND spree_tokenized_permissions.permissable_type = 'Spree::Order'"