+++ core/lib/spree/testing_support/controller_requests.rb
@@ -77,5 +77,3 @@ module Spree
     end
   end
 end
-
-