+++ frontend/spec/controllers/spree/current_order_tracking_spec.rb
@@ -12,8 +12,8 @@ describe 'current order tracking' do
   let(:order) { FactoryGirl.create(:order) }
 
   it 'automatically tracks IP when current_order is called' do
-    request.cookie_jar.signed[:order_id] = order.id
-    get :index, {}, { :order_id => order.id }
+    @request.cookie_jar.signed[:guest_token] = order.guest_token
+    get :index
     controller.current_order.last_ip_address.should == "0.0.0.0"
   end